Feature/okazaki/init

+++ app-manage/src/main/java/jp/co/isol/manage/service/CalcService.java
@@ -1,5 +1,7 @@
 package jp.co.isol.manage.service;
 
+import java.math.BigDecimal;
+
 /**
  * @author kou1210hei<br>
  * 計算サービスインターフェイス
@@ -13,14 +15,14 @@ public interface CalcService {
 	 * @param weight
 	 * @return BMIを計算(小数第2位を四捨五入する)
 	 */
-	public double calcBmi(double height, double weight);
+	public BigDecimal calcBmi(BigDecimal height, BigDecimal weight);
 
 	/**
 	 * 標準体重を計算(小数第2位を四捨五入する)
 	 * @param height
 	 * @return 標準体重を計算(小数第2位を四捨五入する)
 	 */
-	public double calcStandardWeight(double height);
+	public BigDecimal calcStandardWeight(BigDecimal height);
 
 	/**
 	 * 最後に入力した体重と今の体重の差を計算
@@ -28,7 +30,7 @@ public interface CalcService {
 	 * @param now 今の体重
 	 * @return 体重の差(小数第2位を四捨五入する)
 	 */
-	public double calcDiffWeight(double before, double now);
+	public BigDecimal calcDiffWeight(BigDecimal before, BigDecimal now);
 
 }
